Move fiat currencies to constant in PriceWindow
The names of the fiat currencies have been removed from FiatCurrencyPopup.InitialValue and moved into PriceWindow.kCurrencyCodes. This constant holds a JSON string describing a dictionary with the ISO 4217 codes as the keys and the currency names as the values. This was necessary in order to get the popup menu's RowTags set to the codes.