Fix bug in CFString constructor with missing encoding
If the string passed to the constructor has no encoding set, it will silently fail, returning an invalid CFString. The problem is that it tried to use  kCFStringEncodingInvalidId as a default encoding, but this is not actually a valid encoding type--rather, it's a return value indicating an existing CFString doesn't have a valid encoding. I've given it a default encoding of kCFStringEncodingASCII as a lowest common denominator.