Turn App.Prefs back into a regular property
I'd made it a computed property while trying to figure out the bug in setting a value, but that didn't turn out to be necessary. Having it a computed property effectively hides the instantiation of the Prefs object, and I prefer that to be in App.Open where it's easy to see and clear that it's part fo the app's setup.