Change PriceWindow.CurrentRates to a dictionary of rates
Instead of storing only a Rates object for the currently selected fiat currency, we'll store a dictionary with the rates for all the currencies. This way when the user changes currencies the new rate is available in the dictionary already and doesn't have to be retrieved, removing the lag that previously existed when changing currencies.