Add method to find the smallest coin
The CoinList.FindSmallestCoin method takes a dictionary with coin weights as the keys (either CoinList.GoldCoins or CoinList.SilverCoins) and returns the weight of the smallest coin as a double. This is used to determine if a fiat price to be met is less than the value of the smallest gold or silver coin.