Add CoinList constructor to create a list with coins
This additional constructor takes a fiat price and a Rates object, and figures out what gold and silver coins match the fiat price. These coins are added to the CoinList object's Coins property. Any remaining weight of silver that's below that of the smallest silver coin is stored in the SilverRemainder property.