Replace disabled close button with AutoQuit
Disabling the close button causes problems on Windows, and we don't want Mac users to close the price window and have the app still running. Since we can't enable or disable the close button in code depending on platform, the best we can do is set App.AutoQuit to true so that it will quit when its last window closes.