Find approximate BTC rate if nothing was retrieved
An obscure currency might have such low volume of Bitcoin trades that there are no rates available, even for thirty days. If that's the case, we'll find an approximate rate by using the gold rate to find the dollars to obscure currency rate, and from there to the bitcoin rate.