Create a Rates method to handle missing BTC rats
This method was created from code that was previously in the GetBitcoinRates method. When called, it sets the bitcoin rate based on the currency to gold to dollars rate and the bitcoin rate for dollars. It's meant to be used if there is no bitcoin rate at all for a currency with low bitcoin trading volume.