Rewrite GetBitcoinRates to use EstimateBitcoinRate
Some code that was previously in this method has been refactored as the new EstimateBitcoinRate method. That method is now being called if there is no JSON value at all for the currency, or if there is one but it doesn't have any 24h, 7d, or 30d rates in it.