Refactor code to set CoinList shared properties
The code to initially set the CoinList.GoldCoins and CoinList.SilverCoins shared properties was being called from the App.Open event handler. That worked, but now that the JSON the properties are set from is a CoinList constant it makes more sense to do this in the constructor.

Now the default (parameterless) constructor checks if the shared properties are nil, and if one of them is it calls the CoinList.SetKnownCoins method to populate them.