Update indexes in control sets for relative values
The indexes need to correspond to their relative values to simplify matching them with the dictionaries in the array properties. For example, the smallest gold coin is a 1/10 ounce coin, so its dictionary in GoldCoinForCheckBox will be at index 0 and therefore the checkbox in the control set has also been moved to index 0.