Add PreferencesWindow properties for default prefs
We need to pass something as a default anytime we retrieve a value from the TTsSmartPreferences object, and we'll have to retrieve an array (instead of modifying it in place) in the Action handlers because of Xojo limits. Since we'll be using the default prefs in both Open and Action event handlers, they should be properties.

Note that the default prefs properties are arrays of Variants, not arrays of Booleans, even though they will hold booleans. This is because Xojo can't cast arrays, and when we retrieve a pref for coins that is already set, we'll get back an array of variants. Even though everything is actually a boolean, we're still forced to convert the variants (even defaults that we just made into variants from booleans) back into booleans. Sometimes Xojo sucks.