Instantiate PreferencesWindow on launch
We want PreferencesWindow to be instantiated on launch by creating one at the end of the App.Open event handler so that the Open event handlers for the checkbox control sets will run, thus setting the coins to be used by CoinList according to the preferences as read on launch. Since PreferencesWindow.Visible is set to False by default, and we're not changing it, the PreferencesWindow object we create in App.Open is never visible to the user.