Fix exception when number of coins has changed
If the number of coins has changed, the lenght of the array of checkbox states retrieved from the prefs will be different than the size of the control, causing an out of bound exception. There's now a check in the open handlers for the control sets to make sure the array size matches the control set size, and if not it replaces the prefs with the defaults.