Move kCoinWeights constant to Constants module
The JSON with the coin information has been moved from the CoinList class to the Constants module. The code to read the JSON will be moving to the Coin class, so I decided it makes more sense to move this out to the Constants module to avoid confusion about where it is.